+++ src/utility_functions.rs
@@ -18,6 +18,21 @@ pub fn fpowi( base:f32, power:u32 ) -> f32 {
     return temp;
 }
 
+/// Numerical square root of an integer.
+pub fn sqrt( item:i32 ) -> i32 {
+    // Initial approximation
+    let mut root:i32 = item/2;
+    let mut y:i32 = 1;
+    // Accuracy level
+    let error:i32 = 1;
+    while  error <= root - y
+    {
+        root = (root + y) / 2;
+        y = item / root;
+    }
+    return root;
+}
+
 #[cfg(test)]
 mod tests {
     #[test]
@@ -27,5 +42,9 @@ mod tests {
     #[test]
     fn fpowi() {
         assert_eq!{super::fpowi(3.0,2), 9.0};
+    }
+    #[test]
+    fn sqrt() {
+        assert_eq!{super::sqrt(100), 10};
     }
 }